requirements.txt
streamlit
supabase
opencv-python-headless
pyzbar
numpy
Pillow